Added __init__.py's and renamed utils to rp_utils
