+++ core/log_curve.py
@@ -6,14 +6,14 @@ Module for handling LogCurve objects
     GNU Lesser General Public License, Version 3
     (https://www.gnu.org/copyleft/lesser.html)
 """
-from utils.attribdict import AttribDict
-from utils.utils import arrange_logging, info
+from rp_utils.attribdict import AttribDict
+from rp_utils.utils import arrange_logging, info
 from datetime import datetime
 from copy import deepcopy
 import numpy as np
 
-from utils.signal_analysis import rolling_window
-from utils.signal_analysis import smooth as _smooth
+from rp_utils.signal_analysis import rolling_window
+from rp_utils.signal_analysis import smooth as _smooth
 
 
 class Header(AttribDict):
