+++ core/well.py
@@ -21,25 +21,26 @@ import numpy as np
 import logging
 import re
 import os
+import sys
 import matplotlib.pyplot as plt
 from matplotlib.font_manager import FontProperties
 
-from utils.attribdict import AttribDict
-from utils.utils import info
-#from utils.utils import nan_corrcoef
-from utils.utils import log_header_to_template as l2tmpl
-from utils.utils import log_table_in_smallcaps as small_log_table
-import utils.io as uio
-from utils.io import convert
-import utils.masks as msks
-from utils.utils import arrange_logging
-from utils.harmonize_logs import harmonize_logs as fixlogs
+from rp_utils.attribdict import AttribDict
+from rp_utils.utils import info
+#from rp_utils.utils import nan_corrcoef
+from rp_utils.utils import log_header_to_template as l2tmpl
+from rp_utils.utils import log_table_in_smallcaps as small_log_table
+import rp_utils.io as uio
+from rp_utils.io import convert
+import rp_utils.masks as msks
+from rp_utils.utils import arrange_logging
+from rp_utils.harmonize_logs import harmonize_logs as fixlogs
 from plotting import crossplot as xp
 from core.minerals import MineralMix
 from core.log_curve import LogCurve
 import rp.rp_core as rp
-from utils.convert_data import convert as cnvrt
-import utils.definitions as ud
+from rp_utils.convert_data import convert as cnvrt
+import rp_utils.definitions as ud
 
 # global variables
 supported_version = {2.0, 3.0}
@@ -566,7 +567,7 @@ class Well(object):
         :param templates:
             dict
             templates that can contain the desired information for wells
-            templates = utils.io.project_tempplates(wp.project_table)
+            templates = rp_utils.io.project_tempplates(wp.project_table)
         :param search_keys:
             list
             List of strings that can be the key for the desired information in the well header
@@ -768,7 +769,7 @@ class Well(object):
             dict
             Dictionary of working intervals
             E.G.
-            wis = utils.io.project_working_intervals(project_table.xlsx)
+            wis = rp_utils.io.project_working_intervals(project_table.xlsx)
         :param method:
             str
             'Voigt' for the upper bound or Voigt average
@@ -933,7 +934,7 @@ class Well(object):
             name of the mask
         :param tops:
             dict
-            as returned from utils.io.read_tops() function
+            as returned from rp_utils.io.read_tops() function
         :param use_tops:
             list
             List of top names inside the tops dictionary that will be used to mask the data
@@ -944,7 +945,7 @@ class Well(object):
             working intervals, as defined in the "Working intervals" sheet of the project table, and
             loaded through:
             wp = Project()
-            wis = utils.io.project_working_intervals(wp.project_table)
+            wis = rp_utils.io.project_working_intervals(wp.project_table)
 
         :param wi_name:
             str
@@ -1186,7 +1187,7 @@ class Well(object):
             boolean numpy array of same length as xdata
         :param tops:
             dict
-            as returned from utils.io.read_tops() function
+            as returned from rp_utils.io.read_tops() function
         :param wis:
             dict
             dictionary of working intervals
@@ -1196,7 +1197,7 @@ class Well(object):
             matplotlib.axes._subplots.AxesSubplot object
         :param templates:
             dict
-            templates dictionary as returned from utils.io.project_templates()
+            templates dictionary as returned from rp_utils.io.project_templates()
         :param savefig:
             str
             full path name of file to save plot to
@@ -1301,7 +1302,7 @@ class Well(object):
             E.G.
             > from well_project import Project
             > wp = Project()
-            > import utils.io as uio
+            > import rp_utils.io as uio
             > well_table = uio.project_wells(wp.project_table, wp.working_dir)
         :param index:
             int
@@ -1917,7 +1918,7 @@ class Block(object):
         Converts sonic to velocity
         :return:
         """
-        from utils.convert_data import convert
+        from rp_utils.convert_data import convert
 
         for ss, vv, vtype in zip(
                 ['ac', 'acs'], ['vp', 'vs'], ['P velocity', 'S velocity']
