+++ rp_utils/calc_stats.py
@@ -4,11 +4,11 @@ import logging
 import os
 import matplotlib.pyplot as plt
 
-import utils.io as uio
+import rp_utils.io as uio
 from plotting import crossplot as xp
-from utils.utils import nan_corrcoef
-from utils.utils import log_table_in_smallcaps as small_log_table
-import utils.definitions as ud
+from rp_utils.utils import nan_corrcoef
+from rp_utils.utils import log_table_in_smallcaps as small_log_table
+import rp_utils.definitions as ud
 
 logger = logging.getLogger(__name__)
 def_msk_name = 'Mask'  # default mask name
@@ -71,7 +71,7 @@ def calc_stats2_tops(
             {'Volume': ['<', 0.5], 'Porosity': ['>', 0.1]}
     :param templates:
         dict
-        templates dictionary as returned from utils.io.project_templates()
+        templates dictionary as returned from rp_utils.io.project_templates()
     :param rokdoc_output:
         str
         full path name of file of which should contain the averages (RokDoc format)
@@ -184,7 +184,7 @@ def calc_stats2(
         working intervals, as defined in the "Working intervals" sheet of the project table, and
         loaded through:
         wp = Project()
-        wis = utils.io.project_working_intervals(wp.project_table)
+        wis = rp_utils.io.project_working_intervals(wp.project_table)
 
     :param wi_names:
         list of working interval names
@@ -197,7 +197,7 @@ def calc_stats2(
             {'Volume': ['<', 0.5], 'Porosity': ['>', 0.1]}
     :param templates:
         dict
-        templates dictionary as returned from utils.io.project_templates()
+        templates dictionary as returned from rp_utils.io.project_templates()
     :param rokdoc_output:
         str
         full path name of file of which should contain the averages (RokDoc format)
