+++ rp_utils/half_space_avo_monte_carlo.py
@@ -17,7 +17,7 @@ from matplotlib.font_manager import FontProperties
 
 import rp.rp_core as rp
 from plotting import crossplot as myxp, plot_reflectivity as mypr
-from utils import curve_fitting as mycf
+from rp_utils import curve_fitting as mycf
 
 msymbols = np.array(['o','s','v','^','<','>','p','*','h','H','+','x','D','d','|','_','.','1','2','3','4','8'])
 cnames = list(np.roll([str(u) for u in colors.cnames.keys()], 10))  # the first 10 elements have poor colors
@@ -48,7 +48,7 @@ def plot_multi_interfaces(sums, intfs, fbase=None, templates=None, suffix=None):
         if None, no plots are saved
     :param templates:
         dict
-        templates dictionary as returned from utils.io.project_templates()
+        templates dictionary as returned from rp_utils.io.project_templates()
     :param suffix:
         str
         Suffix added to output plots (png) to ease separating output from eachother
