+++ unit_tests/test_calc_stats.py
@@ -1,6 +1,6 @@
 import unittest
 from core.well import Project
-import utils.calc_stats as uc
+import rp_utils.calc_stats as uc
 
 def_lb_name = 'Logs'  # default Block name
 def_msk_name = 'Mask'  # default mask name
