+++ unit_tests/test_masking.py
@@ -1,5 +1,5 @@
 import unittest
-import utils.io as uio
+import rp_utils.io as uio
 import os
 from core.well import Project
 from core.well import Well
